requirements.txt
telethon
aiogram
python-dotenv
requests
pillow
apscheduler